include module_visibility in core
