+++ app/cruds/cruds_module_visibility.py
@@ -4,7 +4,7 @@ from typing import Sequence
 from sqlalchemy import delete, select
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from app.models import models_core, models_module_visibility
+from app.models import models_core
 
 
 async def get_modules_by_user(
@@ -16,11 +16,9 @@ async def get_modules_by_user(
     userGroupIds = list(map(lambda group: group.id, user.groups))
 
     result = await db.execute(
-        select(models_module_visibility.ModuleVisibility.root)
-        .where(
-            models_module_visibility.ModuleVisibility.allowed_group_id.in_(userGroupIds)
-        )
-        .group_by(models_module_visibility.ModuleVisibility.root)
+        select(models_core.ModuleVisibility.root)
+        .where(models_core.ModuleVisibility.allowed_group_id.in_(userGroupIds))
+        .group_by(models_core.ModuleVisibility.root)
     )
 
     return result.unique().scalars().all()
@@ -34,8 +32,8 @@ async def get_allowed_groups_by_root(
 
     result = await db.execute(
         select(
-            models_module_visibility.ModuleVisibility.allowed_group_id,
-        ).where(models_module_visibility.ModuleVisibility.root == root)
+            models_core.ModuleVisibility.allowed_group_id,
+        ).where(models_core.ModuleVisibility.root == root)
     )
 
     resultList = result.unique().scalars().all()
@@ -50,22 +48,22 @@ async def get_module_visibility(
     root: str,
     group_id: str,
     db: AsyncSession,
-) -> models_module_visibility.ModuleVisibility | None:
+) -> models_core.ModuleVisibility | None:
     """Return module visibility by root and group id"""
 
     result = await db.execute(
-        select(models_module_visibility.ModuleVisibility).where(
-            (models_module_visibility.ModuleVisibility.allowed_group_id == group_id)
-            & (models_module_visibility.ModuleVisibility.root == root)
+        select(models_core.ModuleVisibility).where(
+            (models_core.ModuleVisibility.allowed_group_id == group_id)
+            & (models_core.ModuleVisibility.root == root)
         )
     )
     return result.unique().scalars().first()
 
 
 async def create_module_visibility(
-    module_visibility: models_module_visibility.ModuleVisibility,
+    module_visibility: models_core.ModuleVisibility,
     db: AsyncSession,
-) -> models_module_visibility.ModuleVisibility:
+) -> models_core.ModuleVisibility:
     """Create a new module visibility in database and return it"""
 
     db.add(module_visibility)
@@ -83,12 +81,9 @@ async def delete_module_visibility(
     db: AsyncSession,
 ):
     await db.execute(
-        delete(models_module_visibility.ModuleVisibility).where(
-            (models_module_visibility.ModuleVisibility.root == root)
-            & (
-                models_module_visibility.ModuleVisibility.allowed_group_id
-                == allowed_group_id
-            )
+        delete(models_core.ModuleVisibility).where(
+            (models_core.ModuleVisibility.root == root)
+            & (models_core.ModuleVisibility.allowed_group_id == allowed_group_id)
         )
     )
     await db.commit()
