+++ app/models/models_core.py
@@ -103,3 +103,10 @@ class CoreGroup(Base):
         secondary="core_membership",
         back_populates="groups",
     )
+
+
+class ModuleVisibility(Base):
+    __tablename__ = "module_visibility"
+
+    root: Mapped[str] = mapped_column(String, primary_key=True)
+    allowed_group_id: Mapped[str] = mapped_column(String, primary_key=True)
