+++ app/schemas/schemas_core.py
@@ -276,3 +276,19 @@ class CoreBatchDeleteMembership(BaseModel):
 class CoreMembershipDelete(BaseModel):
     user_id: str
     group_id: str
+
+
+class ModuleVisibility(BaseModel):
+    root: str
+    allowed_group_ids: list[str]
+
+    class Config:
+        orm_mode = True
+
+
+class ModuleVisibilityCreate(BaseModel):
+    root: str
+    allowed_group_id: str
+
+    class Config:
+        orm_mode = True
