--- app/schemas/schemas_module_visibility.py
@@ -1,17 +1,0 @@
-from pydantic import BaseModel
-
-
-class ModuleVisibility(BaseModel):
-    root: str
-    allowed_group_ids: list[str]
-
-    class Config:
-        orm_mode = True
-
-
-class ModuleVisibilityCreate(BaseModel):
-    root: str
-    allowed_group_id: str
-
-    class Config:
-        orm_mode = True
